restructure notebook to skip certain cells
